+++ CHANGELOG.md
@@ -11,6 +11,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - Warn before sending emails with an empty subject with `empty-subject-warning`
   in `aerc.conf`.
 - IMAP now uses the delimiter advertised by the server
+- Completions for `:mkdir`
 
 ### Fixed